Some warning corrections
char * string literals and 64-to-32-bit implicit conversions.

+++ source/textscreen/txt_fileselect.c
@@ -89,7 +89,7 @@ static char *ExecReadOutput(char **argv)
             completed = 1;
         }
 
-        bytes = read(pipefd[0], buf, sizeof(buf));
+        bytes = (int)read(pipefd[0], buf, sizeof(buf));
 
         if (bytes < 0)
         {
